restructure binaries
