+++ Cargo.toml
@@ -9,7 +9,8 @@ members = [
   "crates/control",
   "crates/filtering",
   "crates/framework",
-  "crates/hulk",
+  "crates/hulk_nao",
+  "crates/hulk_webots",
   "crates/kinematics",
   "crates/nao",
   "crates/nao_camera",
