--- crates/hulk/src/cyclers.rs
@@ -1,5 +1,0 @@
-#![allow(clippy::too_many_arguments)]
-#![allow(clippy::redundant_clone)]
-#![allow(clippy::needless_question_mark)]
-#![allow(clippy::nonminimal_bool)]
-include!(concat!(env!("OUT_DIR"), "/cyclers.rs"));
