--- crates/hulk/src/nao/mod.rs
@@ -1,8 +1,0 @@
-mod camera;
-mod double_buffered_reader;
-mod hula;
-mod hula_wrapper;
-mod interface;
-mod microphones;
-
-pub use interface::Interface;
