--- crates/hulk/src/perception_databases.rs
@@ -1,1 +1,0 @@
-include!(concat!(env!("OUT_DIR"), "/perception_databases.rs"));
