--- crates/hulk/src/structs.rs
@@ -1,1 +1,0 @@
-include!(concat!(env!("OUT_DIR"), "/structs.rs"));
