--- crates/hulk/src/webots/mod.rs
@@ -1,9 +1,0 @@
-mod camera;
-mod force_sensitive_resistor_devices;
-mod interface;
-mod intertial_measurement_unit_devices;
-mod joint_devices;
-mod keyboard_device;
-mod sonar_sensor_devices;
-
-pub use interface::Interface;
