+++ crates/hulk_nao/src/camera.rs
