+++ crates/hulk_nao/src/double_buffered_reader.rs
