+++ crates/hulk_webots/Cargo.toml
@@ -1,0 +1,42 @@
+[package]
+name = "hulk_webots"
+version = "0.1.0"
+edition = "2021"
+license = "GPL-3.0-only"
+homepage = "https://github.com/hulks/hulk"
+
+[dependencies]
+audio = { workspace = true }
+chrono = { workspace = true }
+color-eyre = { workspace = true }
+communication = { workspace = true, features = ["server"] }
+control = { workspace = true }
+ctrlc = { workspace = true }
+fern = { workspace = true }
+framework = { workspace = true }
+i2cdev = { workspace = true }
+ittapi = {  workspace = true }
+log = { workspace = true }
+nalgebra = { workspace = true }
+parking_lot = { workspace = true }
+rand = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+serialize_hierarchy = { workspace = true }
+spl_network = { workspace = true }
+spl_network_messages = { workspace = true }
+tokio = { workspace = true }
+tokio-util = { workspace = true }
+types = { workspace = true }
+vision = { workspace = true }
+webots = { workspace = true }
+
+[build-dependencies]
+code_generation = { workspace = true }
+color-eyre = { workspace = true }
+convert_case = { workspace = true }
+proc-macro2 = { workspace = true }
+quote = { workspace = true }
+serde = { workspace = true }
+source_analyzer = { workspace = true }
+toml = { workspace = true }
