+++ crates/hulk_webots/framework.toml
@@ -1,0 +1,90 @@
+[[cyclers]]
+name = "Vision"
+kind = "Perception"
+instances = ["Top", "Bottom"]
+module = "vision"
+nodes = [
+  { module = "image_receiver", is_setup = true },
+  "ball_detection",
+  "camera_matrix_extractor",
+  "feet_detection",
+  "field_border_detection",
+  "field_color_detection",
+  "image_segmenter",
+  "limb_projector",
+  "line_detection",
+  "perspective_grid_candidates_provider",
+  "robot_detection",
+  "segment_filter",
+]
+
+[[cyclers]]
+name = "Control"
+kind = "RealTime"
+module = "control"
+nodes = [
+  { module = "sensor_data_receiver", is_setup = true },
+  "active_vision",
+  "ball_filter",
+  "ball_state_composer",
+  "behavior::node",
+  "button_filter",
+  "camera_matrix_calculator",
+  "center_of_mass_provider",
+  "fall_state_estimation",
+  "game_controller_filter",
+  "game_state_filter",
+  "ground_contact_detector",
+  "ground_provider",
+  "kick_selector",
+  "kinematics_provider",
+  "led_status",
+  "localization",
+  "motion::arms_up_squat",
+  "motion::condition_input_provider",
+  "motion::dispatching_interpolator",
+  "motion::energy_saving_stand",
+  "motion::fall_protector",
+  "motion::head_motion",
+  "motion::joint_command_sender",
+  "motion::jump_left",
+  "motion::jump_right",
+  "motion::look_around",
+  "motion::look_at",
+  "motion::motion_selector",
+  "motion::sit_down",
+  "motion::stand_up_back",
+  "motion::stand_up_front",
+  "motion::step_planner",
+  "motion::walk_manager",
+  "motion::walking_engine",
+  "obstacle_filter",
+  "odometry",
+  "orientation_filter",
+  "penalty_shot_direction_estimation",
+  "primary_state_filter",
+  "role_assignment",
+  "rule_obstacle_composer",
+  "sole_pressure_filter",
+  "sonar_filter",
+  "support_foot_estimation",
+  "whistle_filter",
+  "world_state_composer",
+]
+
+[[cyclers]]
+name = "SplNetwork"
+kind = "Perception"
+module = "spl_network"
+nodes = [
+  { module = "message_receiver", is_setup = true },
+]
+
+[[cyclers]]
+name = "Audio"
+kind = "Perception"
+module = "audio"
+nodes = [
+  { module = "microphone_recorder", is_setup = true },
+  "whistle_detection",
+]
